Implement BlockTokenizer.get_rest_of_line()
This returns the rest of a line as plain text without any associated
token type.

It is useful for structures that cause the token type to be ignored. For
example in a code block, a header will not be rendered as a header but as a # followed by a
line:

    # not a header