Implement open-close block parsing
The block parser now includes a open_block_stack which will keep track
of which blocks are currently "open" (can have children added). This way
instead of explicitly stating which block to add a child to, the parser
just says to add the child to the lowest open block on the stack.

There were also some changes to the dom structure included:

Separate new CodeBlock lines with '\n' instead of with separate <code>
tags

Wrap all content inside of a <body> tag inside <html>